done all components
