+++ website/src/routes/components/monet-inline.svelte
@@ -1,10 +1,10 @@
 <script>
+	import { onMount } from 'svelte';
 	import Page from '$lib/components/components/Page.svelte';
 	import RadioGroup from '$lib/components/components/RadioGroup.svelte';
 	import Text from '$lib/components/components/Text.svelte';
 	import userPref from '$lib/components/components/monetization-type';
 	import { defaults as _defaults } from '@loremlabs/monet/dist/components/monet-inline/app.mjs';
-	import { onMount } from 'svelte';
 	/** @type {import("@loremlabs/monet/src/components/MonetInline/defaults")} */
 	const defaults = _defaults;
